Support integer values bigger than Int32 in min() and max() functions (#31)
* Support integer values bigger than Int32 in min() and max() functions

* Add regression test: number_int64.json to support integer values bigger than Int32 for min() and max() functions

+++ src/jmespath.net/Functions/MaxFunction.cs
@@ -30,7 +30,7 @@ namespace DevLab.JmesPath.Functions
                             return GetMax<double>(array);
 
                         else /* if (token.Type == JTokenType.Integer) */
-                            return GetMax<int>(array);
+                            return GetMax<long>(array);
                     }
 
                 case "string":
